Renames EmbeddingService -> embedding-service for consistency.
