+++ src/services/embedding-service.ts
